app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
gunicorn==21.2.0
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
pytz
redis
celery
djangorestframework
djoser
django-cors-headers
pillow
channels
channels_redis==3.3.1